+++ packages/reactabular-highlight/src/index.js
@@ -1,3 +1,0 @@
-export { default as cell } from './cell';
-export { default as value } from './value';
-export { default as highlighter } from './highlighter';