+++ packages/reactabular/src/index.js
@@ -1,6 +1,5 @@
 import * as Table from 'reactabular-table';
 import * as Sticky from 'reactabular-sticky';
-import * as highlight from 'reactabular-highlight';
 import * as resolve from 'reactabular-resolve';
 import SearchColumns from 'reactabular-search-columns';
 import Search from 'reactabular-search-field';
@@ -10,7 +9,6 @@ export {
   Sticky,
   SearchColumns,
   Search,
-  highlight,
   resolve
 };